CSU06: implement backend to show reports and delete users and reports

+++ backend/src/main/java/com/campimove/backend/repositories/UserReportRepository.java
@@ -3,5 +3,5 @@ package com.campimove.backend.repositories;
 import com.campimove.backend.entities.UserReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserReportRepository extends JpaRepository<UserReport,Integer> {
+public interface UserReportRepository extends JpaRepository<UserReport,Long> {
 }
